Change already visited submission logic to use sid
sid is a base36 int. We convert that sid into a python int and
check if the new submissions' sid is less than or equal to the
last sid saved. If it is, we have already seen the post. Stop
processing. The reason we check if it is less than (and the
reason that we can't compare sid strings) is that a submission
can be deleted, and the next time we check it, it can't be found.